refactor: (#849) VoteDTO 통합
