--- dms-core/src/main/kotlin/team/aliens/dms/domain/vote/dto/response/ModelStudentResponse.kt
@@ -1,11 +1,0 @@
-package team.aliens.dms.domain.vote.dto.response
-
-import java.util.UUID
-
-data class ModelStudentResponse(
-    val id: UUID,
-    val gcn: String,
-    val name: String,
-    val profileImageUrl: String?
-
-)
