+++ dms-core/src/main/kotlin/team/aliens/dms/domain/vote/dto/response/VoteResponse.kt
@@ -1,6 +1,68 @@
 package team.aliens.dms.domain.vote.dto.response
 
+import team.aliens.dms.domain.vote.model.VoteType
+import team.aliens.dms.domain.vote.model.VotingTopic
+import java.time.LocalDateTime
 import java.util.UUID
+
+data class VotingTopicResponse(
+    val id: UUID,
+    val topicName: String,
+    val description: String? = null,
+    val startTime: LocalDateTime,
+    val endTime: LocalDateTime,
+    val voteType: VoteType,
+) {
+    companion object {
+        fun from(votingTopic: VotingTopic): VotingTopicResponse {
+            return VotingTopicResponse(
+                id = votingTopic.id,
+                topicName = votingTopic.topicName,
+                description = votingTopic.description,
+                startTime = votingTopic.startTime,
+                endTime = votingTopic.endTime,
+                voteType = votingTopic.voteType,
+            )
+        }
+    }
+}
+
+data class VotingTopicsResponse(
+    val votingTopics: List<VotingTopicResponse>
+)
+
+data class ExcludedStudentResponse(
+    val studentId: UUID,
+    val gcn: String,
+    val name: String
+) {
+    companion object {
+        fun of(studentId: UUID, gcn: String, name: String): ExcludedStudentResponse {
+            return ExcludedStudentResponse(
+                studentId = studentId,
+                gcn = gcn,
+                name = name
+            )
+        }
+    }
+}
+
+data class ExcludedStudentsResponses(
+    val excludedStudents: List<ExcludedStudentResponse>
+)
+
+data class ModelStudentResponse(
+    val id: UUID,
+    val gcn: String,
+    val name: String,
+    val profileImageUrl: String?
+
+)
+
+data class ModelStudentsResponse(
+    val students: List<ModelStudentResponse>
+
+)
 
 data class VoteResponse(
     val id: UUID,
@@ -19,6 +81,45 @@ data class VoteResponse(
             name = name,
             votes = votes,
             classNumber = classNumber
+        )
+    }
+}
+
+data class VotesResponse(
+    val votes: List<VoteResponse>
+) {
+    companion object {
+        fun of(
+            votes: List<VoteResponse>
+        ) = VotesResponse(
+            votes = votes
+        )
+    }
+}
+
+class VotingOption(
+    val id: UUID,
+    val votingOptionName: String,
+) {
+    companion object {
+        fun of(
+            id: UUID,
+            votingOptionName: String
+        ) = VotingOption(
+            id = id,
+            votingOptionName = votingOptionName
+        )
+    }
+}
+
+class VotingOptionsResponse(
+    val votingOptions: List<VotingOption>
+) {
+    companion object {
+        fun of(
+            votingOptions: List<VotingOption>
+        ) = VotingOptionsResponse(
+            votingOptions = votingOptions
         )
     }
 }
