--- dms-core/src/main/kotlin/team/aliens/dms/domain/vote/dto/response/VotesResponse.kt
@@ -1,13 +1,0 @@
-package team.aliens.dms.domain.vote.dto.response
-
-data class VotesResponse(
-    val votes: List<VoteResponse>
-) {
-    companion object {
-        fun of(
-            votes: List<VoteResponse>
-        ) = VotesResponse(
-            votes = votes
-        )
-    }
-}
