--- dms-core/src/main/kotlin/team/aliens/dms/domain/vote/dto/response/VotingOption.kt
@@ -1,18 +1,0 @@
-package team.aliens.dms.domain.vote.dto.response
-
-import java.util.UUID
-
-class VotingOption(
-    val id: UUID,
-    val votingOptionName: String,
-) {
-    companion object {
-        fun of(
-            id: UUID,
-            votingOptionName: String
-        ) = VotingOption(
-            id = id,
-            votingOptionName = votingOptionName
-        )
-    }
-}
