--- dms-core/src/main/kotlin/team/aliens/dms/domain/vote/dto/response/VotingOptionsResponse.kt
@@ -1,13 +1,0 @@
-package team.aliens.dms.domain.vote.dto.response
-
-class VotingOptionsResponse(
-    val votingOptions: List<VotingOption>
-) {
-    companion object {
-        fun of(
-            votingOptions: List<VotingOption>
-        ) = VotingOptionsResponse(
-            votingOptions = votingOptions
-        )
-    }
-}
